Restructure app definition files and code
